requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pymilvus==2.5.0
google-generativeai==0.8.3
google-cloud-secret-manager==2.21.1
pymupdf==1.25.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
streamlit==1.51.0